`string.Template` instead of `str.replace`
Use `string.Template` instead of `str.replace` to process string
substitutions specified with option --dirnames. The motivation is to
allow the substitution of a variable which begins by a substring which
is another variable to be replaced. For example: `build_dir` and
`build_dir_16x12x39` (for LMDZE). Using `string.Template` has the
additional benefits of being more concise to write, probably more
efficient (the line may be processed in a single pass) and clearer in
the input JSON file (no dollar sign in the input JSON file).