No exit after a difference in `nccmp_pism.py`
Use argsparse instead of getopt in `nccmp_pism.py`.

Do not exit after finding a difference in `nccmp_pism.py`. We want to
see the differences for all variables.

No need for default value of argument relative of function
`compare_vars` of `nccmp_pism.py`: it is always called with the actual
argument present.

In `nccmp_pism.py`, no need to print again tol for each variable: it
is always the same.

In `nccmp_pism.py`, clarify description.