Redefine `nccmp.py`
Rewrite `nccmp.py` so that it only describes qualitatively the difference
on data (not meta-data), but does not compute any difference. That is: it
only says which variables differ, and what type of difference (shape,
content...). Computing quantitative difference is rather the job of
`Max_diff_nc` for now. The tolerance option of `nccmp.py` was not used in
`selective_diff.sh` anyway. Add options to stop at the first difference
and to be silent. `nccmp.py` can then be used with `nccmp_meta.py` in lieu
of cmp to recognize NetCDF files which are mysteriously not identical but
have same data and meta-data content.