Move the logic into the initialization of instance
The problem was that it was complicated to bind a function to an
instance (make it a method) by passing the functions as arguments. See
for example [Adding a Method to an Existing Object
Instance](https://stackoverflow.com/questions/972/adding-a-method-to-an-existing-object-instance),
answer from Jason Pratt, or [Run-time method patching in
Python](https://tryolabs.com/blog/2013/07/05/run-time-method-patching-python).