Do not compare again, even if different
Before this commit, we created, for each new run identical to the old
one, an empty file named `identical`. Then the comparison would not be
done again in function `compare`. But if the new run was different
from the old one, the comparison could be done again in function
`compare`. We want the comparison not to be done again if it has
already been done, regardless of the result of the comparison. So
instead of creating an empty file `identical`, we read or write the
return code of the comparison in `identical`.