Read a dictionary of runs
Big change. Read a dictionary of dictionaries instead of a list of
dictionaries. The title becomes the key in the dictionary of runs. The
title was the only compulsory key in the run description. Using a
dictionary of runs is useful to derive new runs from existing runs, as
in the script to create release tests for LMDZ.