Bug fix: pass string argument to subprocess
The bug comes from commit 58a28819. We keep the default value of
tolerance as float, for clarity, even though we only use it to
transmit it to subprocess. Then we also change the type of the command
line argument tolerance to float: this allows argparse to check that
the tolerance is a float value.