Extract code to new script `re_compare.py`
Replace option `re_compar` of `test_compare.py` by new script
`re_compare.py`. The difference is that `re_compare.py` does not try
to archive the test directory in `old_dir`. The motivation is that we
want to enrich the options of `re_compare.py` with options of
`selective_diff.py`.